Update ClientFilter tests
